Working final version of planner

+++ 16782-HW1/code/planner.cpp
@@ -19,10 +19,10 @@
 
 #define NUMOFDIRS 8
 
-void planner_greedy(double *map, int collision_thresh, int x_size, int y_size,
+void planner_greedy(int *map, int collision_thresh, int x_size, int y_size,
                     int robotposeX, int robotposeY, int target_steps,
-                    double *target_traj, int targetposeX, int targetposeY,
-                    int curr_time, double *action_ptr) {
+                    int *target_traj, int targetposeX, int targetposeY,
+                    int curr_time, int *action_ptr) {
   // 8-connected grid
   int dX[NUMOFDIRS] = {-1, -1, -1, 0, 0, 1, 1, 1};
   int dY[NUMOFDIRS] = {-1, 0, 1, -1, 1, -1, 0, 1};
@@ -30,8 +30,8 @@ void planner_greedy(double *map, int collision_thresh, int x_size, int y_size,
   // for now greedily move towards the final target position,
   // but this is where you can put your planner
 
-  int goalposeX = (int)target_traj[target_steps - 1];
-  int goalposeY = (int)target_traj[target_steps - 1 + target_steps];
+  int goalposeX = target_traj[target_steps - 1];
+  int goalposeY = target_traj[target_steps - 1 + target_steps];
   // printf("robot: %d %d;\n", robotposeX, robotposeY);
   // printf("goal: %d %d;\n", goalposeX, goalposeY);
 
@@ -46,8 +46,8 @@ void planner_greedy(double *map, int collision_thresh, int x_size, int y_size,
     int newy = robotposeY + dY[dir];
 
     if (newx >= 1 && newx <= x_size && newy >= 1 && newy <= y_size) {
-      if (((int)map[GETMAPINDEX(newx, newy, x_size, y_size)] >= 0) &&
-          ((int)map[GETMAPINDEX(newx, newy, x_size, y_size)] <
+      if ((map[GETMAPINDEX(newx, newy, x_size, y_size)] >= 0) &&
+          (map[GETMAPINDEX(newx, newy, x_size, y_size)] <
            collision_thresh)) // if free
       {
         disttotarget = (double)sqrt(((newx - goalposeX) * (newx - goalposeX) +
@@ -68,10 +68,9 @@ void planner_greedy(double *map, int collision_thresh, int x_size, int y_size,
   return;
 }
 
-void planner(double *map, int collision_thresh, int x_size, int y_size,
-             int robotposeX, int robotposeY, int target_steps,
-             double *target_traj, int targetposeX, int targetposeY,
-             int curr_time, double *action_ptr) {
+void planner(int *map, int collision_thresh, int x_size, int y_size,
+             int robotposeX, int robotposeY, int target_steps, int *target_traj,
+             int targetposeX, int targetposeY, int curr_time, int *action_ptr) {
 
   static RealTimePlanner rtplanner(map, collision_thresh, x_size, y_size,
                                    target_steps, target_traj, robotposeX,
